feat: add comment editing and pwnership validation

+++ internal/services/comment_service.go
@@ -1,10 +1,13 @@
 package services
 
 import (
+	"errors"
+
 	"github.com/AsterOzlob/content_managment_api/internal/database/models"
 	"github.com/AsterOzlob/content_managment_api/internal/database/repositories"
 	"github.com/AsterOzlob/content_managment_api/internal/dto"
 	logger "github.com/AsterOzlob/content_managment_api/internal/logger"
+	"github.com/AsterOzlob/content_managment_api/pkg/utils"
 )
 
 // CommentService предоставляет методы для управления комментариями.
@@ -19,13 +22,13 @@ func NewCommentService(repo *repositories.CommentRepository, logger logger.Logge
 }
 
 // AddCommentToArticle добавляет комментарий к статье.
-func (s *CommentService) AddCommentToArticle(articleID uint, input dto.CommentInput) (*models.Comment, error) {
+func (s *CommentService) AddCommentToArticle(articleID uint, input dto.CommentInput, userID uint) (*models.Comment, error) {
 	s.Logger.WithField("article_id", articleID).Info("Adding comment to article")
 
 	comment := &models.Comment{
 		ParentID:  input.ParentID,
 		ArticleID: articleID,
-		AuthorID:  input.AuthorID,
+		AuthorID:  userID,
 		Text:      input.Text,
 	}
 
@@ -50,9 +53,42 @@ func (s *CommentService) GetCommentsByArticleID(articleID uint) ([]*models.Comme
 	return comments, nil
 }
 
+// UpdareComment редактирует содержимое комментария.
+func (s *CommentService) UpdateComment(id uint, input dto.CommentInput, userID uint, roles []string) (*models.Comment, error) {
+	s.Logger.WithField("comment_id", id).Info("Updating comment")
+
+	comment, err := s.repo.GetByID(id)
+	if err != nil {
+		return nil, err
+	}
+
+	if !utils.IsOwner(comment.AuthorID, userID, roles) {
+		return nil, errors.New("access denied: you are not the owner or don't have required role")
+	}
+
+	comment.Text = input.Text
+
+	if err := s.repo.Update(comment); err != nil {
+		s.Logger.WithError(err).Error("Failed to update comment in repository")
+		return nil, err
+	}
+
+	return comment, nil
+}
+
 // DeleteComment удаляет комментарий по ID.
-func (s *CommentService) DeleteComment(commentID uint) error {
+func (s *CommentService) DeleteComment(commentID uint, userID uint, userRoles []string) error {
 	s.Logger.WithField("comment_id", commentID).Info("Deleting comment in service")
+
+	comment, err := s.repo.GetByID(commentID)
+	if err != nil {
+		return err
+	}
+
+	// Проверяем права через IsOwner
+	if !utils.IsOwner(comment.AuthorID, userID, userRoles) {
+		return errors.New("access denied: you are not the owner or don't have required role")
+	}
 
 	if err := s.repo.Delete(commentID); err != nil {
 		s.Logger.WithError(err).Error("Failed to delete comment from repository")
